Cortex: fix the catalog location of the main application
Previously, hrev55011 introduced localization for Cortex, which then
prompted to close #7530. It has then been reopened, when it turns out
that the translations for the container application (RouteApp) did not
work, whereas the individual add-ons/modules were translated.

The cause is that by default BCatalog looks up the translations based on
the subtype part of the signature. This is x-vnd.Cortex.Route (without
the application/ supertype). This change will place the translations in
the right place of the file system.

The add-ons were never affected, since they BCatalog is explicitly told
to find the translations for the entire signature, like:

static BCatalog sCatalog("application/x-vnd.Cortex.InfoView");

Even so, it was chosen to omit the `application` supertype from the
signature for the shared code as well.

This should fix #7530 for good.

Change-Id: Iff18fabef7aba68602e49db1e98cfed2f486f545