Update patch set 5
Patch Set 5:

(2 comments)

A agree all the FDT stuff should be a real bus.

The reason for the support code's existence like this however is to allow bootloader platforms like u-boot to get access to a serial uart via serial uart drivers "included" in the bootloader.

By accepting this, we would be committing to rely on bios services for uart / debug bus discovery on all platforms / architectures.

See:
  src/system/boot/platform/u-boot/serial.cpp:#include "fdt_serial.h"
  src/system/kernel/arch/ppc/arch_platform.cpp:#include "fdt_serial.h"

Including Francois since it is relevant to things he's worked on.

Patch-set: 5
CC: Gerrit User 1000022 <1000022@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
CC: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>