Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Shouldn't we rather fix the MIME type instead?

I am slightly on the fence here. The fix for consistency would actually to drop the 'application/' supertype from all the add-on modules, as this is consistently omitted in our other localized add-ons. The documentation for BCatalog speaks about the application signature, which would actually include the 'application/' supertype.

This is the practical fix, which makes the Cortex host application find its translations, but leaves the add-ons untouched.

Patch-set: 1