Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Could we push the fields first, and then have them filled when we have enough information from other OS behavior?

It still feels like we are compensating for poorly written, non-portable software. As for LLVM, I have actually gotten a patch accepted that works around it (should land in LLVM 13).

As for the way this is done now: common usage is to create an uninitialized object of the rusage struct, and rely on the get_rusage() call to initialize it fully. This approach will lead to only the first two fields being initialized, and the rest being filled with bogus data.

Patch-set: 2
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>