Update patch set 2
Patch Set 2:

> > Patch Set 2:
> > 
> > Could we push the fields first, and then have them filled when we have enough information from other OS behavior?
> 
> It still feels like we are compensating for poorly written, non-portable software. As for LLVM, I have actually gotten a patch accepted that works around it (should land in LLVM 13).

We already compensate by patching at HaikuPorts, and have to maintain most of these patches for ever.

> 
> As for the way this is done now: common usage is to create an uninitialized object of the rusage struct, and rely on the get_rusage() call to initialize it fully. This approach will lead to only the first two fields being initialized, and the rest being filled with bogus data.

Sure we can fill the fields with zero and add the compatibility symbol. But if fields with zero won't be accepted (see what PulkoMandy wrote), it's a wasted effort. And anyway these fields won't be accessed because programs which might use them are patched at HaikuPorts to not use these fields...

Patch-set: 2