Update patch set 2
Patch Set 2:

> Sure we can fill the fields with zero and add the compatibility symbol. But if fields with zero won't be accepted (see what PulkoMandy wrote), it's a wasted effort. And anyway these fields won't be accessed because programs which might use them are patched at HaikuPorts to not use these fields...

Setting them to 0 is better than not initializing them at all, and provides us with an easy upgrade path when we implement some of these fields. To do this we need to set up the symbol versionning so that old applications will still work.

That seems reasonable: it is better to do this once, than introduce a new symbol version everytime we add a field to the structure.

Patch-set: 2