Update patch set 4
Patch Set 4:

What I checked: the assembly of wait4_current is correct (8 longs will be zeroed with rep sto), tested with wait4_test unit test.

Versions in BeBuild.h are not related with symbol versioning. Here we introduce a new symbol, which will be used by compilers from now on. The old symbol stays available, beta packages continue to work and that's the point.

What's missing is the regeneration of the libroot stubs, but a few updates are missing, so it can be done in a later change.

Patch-set: 4