Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> Looks like there is quite some work left with clang-format :-)

Yes, the "good" parts of the changes were already merged, this change request is here to help track and review the work to be done on clang-format.

In this latest version Saloni tried to add automatic indentation in BLayoutBuilder use. Some other issues are already known as well. When they are all fixed we can move on to formatting other parts of the code as well.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>