Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi Adrien; I see your comment here about the review being difficult.  I would like to get this through as it is a step in getting some problems sorted out in the application.  I won't merge until after the R3 branch, but do you object to me merging anyway?

No objections from me, reviews are not mandatory in Haiku (they are only recommended), if you think it's ready, you can merge it. Or we can wait until someone takes the time to review it more in depth.

In my case, with the upcoming beta and GSoC mentoring going on, I'm not sure if I will have time to take a closer look.

Patch-set: 2