Update patch set 1
Patch Set 1:

In need of additional feedback.

The goal is to use a dynamically allocated array and calculate the string length using a while loop until the next value is equal to NULL, then slightly adjust the existing code in order to affect the logic the least amount possible. Doing all of this, however, requires getting rid of sizeof(), which renders the condition using it quite... fragile, as far as figuring out an equivalent condition that will work with dynamic strings is concerned.

I thought that an empty string would mean that it would be equal to NULL, although, considering the very next lines of the code, I am worried that the string may actually not be capable of being empty, in the sense that the newline could be there too. The condition, however, is also confusing in the same regard, and I am really not sure as to how I could test this in a conclusive manner.

P.S. Another goal is to allow input exceeding 126 characters.

Patch-set: 1