Update patch set 1
Patch Set 1: Code-Review-2

This function returns a pointer to a static buffer. It is not possible to use dynamic allocation here, because you can't know when it is safe to delete or realloc the buffer, because the application which called this is allowed to reuse the pointer (and, in fact, the documentation recommends that the app erases the buffer itself). You would also not be able to handle memory allocation failures correctly.

Please leave this function as it is. If you need something that supports larger buffers, consider implementing readpassphrase() which was introduced in freebsd to overcome the limitations of this API

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2