Update patch set 1
Patch Set 1:

Not sure if this deserves to get removed, but I made this change since I don't think I'm supposed to push this when making any sort of change, because of the fact that I have to be very careful so that I won't push this file remotely and because of the fact that this file does not exist upstream. That makes me think that it's a leftover artifact.

Of course, I will promptly add that to my personal global gitignore file, but I'm assuming that I'm not the only one with this issue.

Patch-set: 1