Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > What's this file and why is it created in the source directory?
> 
> I am not very sure, honestly, but I found a mention of it here: https://github.com/haiku/haiku/blob/a182bd6eb5656b9fa14556fd0006bd514c6516fe/src/system/boot/Jamfile#L310
> 
> Hope this is enough.

If it is generated by the Haiku build, it should be moved into the generated directory (using a MakeLocate in the jam rule that creates it, for example). No files should be generated outside of the generated directory.

If it is meant to be created by the user, it should be placed in some directory (build/jam next to UserBuildConfig?) and there should be some documentation about it.

It looks like you are attempting to hide a bug instead of fixing it. Please don't do that here 😊

Patch-set: 1