Update patch set 2
Patch Set 2:

Wait, are you running jam in the top level, or in the generated directory?

There's a chance that the combination of running ../configure from the clean build directory (as described in https://www.haiku-os.org/guides/building/compiling-x86_64) also means that efi.map (and nothing else) will be put in the top level directory, if the developer runs jam from the top level directory instead of the "generated" folder.

Patch-set: 2