Update patch set 1
Patch Set 1:

yeah it's feature addition but it's enhanced on something already tested: the added keycode is specific, so it can't hurt anything else. the fn-lock is checked against a presence bit on a WMI specific key. it's low risk and nice to have probably for some vivobook owners, but won't hurt anyone if not cherry-picked.

Patch-set: 1