Update patch set 1
Patch Set 1:

Please note that this is a rewrite and more recent versions of the file (both FreeBSD/OpenBSD), application-specific forks of this particular file and man pages were used as a point of reference, instead of the relatively old version that currently exists on the Haiku tree.

The goal is to follow the pattern of using our own implementations for *BSD compatibility, while making sure that the said implementation will still work in the presence of asynchronous events and coroutines running in the background. For example, the assertion will make it easier to deal with programs that use incompatible signals as far as porting is concerned, instead of having to deal with "mysterious" undefined behavior.

Patch-set: 1