headers/posix: Define OpenBSD's _NSIG
Although NSIG is defined in FreeBSD, this variable is not universal.

_NSIG tends to be used more widely across BSD-based/BSD-inspired kernels, and
FreeBSD even directly compensates for that by defining _NSIG as NSIG. Such
examples include OpenBSD and NetBSD, which internally use _NSIG and have added
support for NSIG in the opposite way that FreeBSD natively supports NSIG
and has added support for _NSIG.

That way, software developers that use that variable do not have to be concerned
with whether their software will run on FreeBSD. However, as things are right
now, that software won't run on Haiku and we are currently arbitrarily
limiting how "compatible" our system is with BSD-based software.

Although NSIG (or _NSIG) are not variables that should generally be used,
we shouldn't dictate why software developers use it and complain if their software
can still run perfectly fine on our platform without causing any major disruptions
that go beyond the introduction of a simple macro. Internal source files for both
operating systems, as well as external dependencies such as gettext and gdb,
utilize either of these variables.

That way, we won't be strictly limiting low-level software that wasn't explicitly
written for FreeBSD.

Change-Id: I3209d2f258699eb7f1a7abb3a86c360b12ddc376