Update patch set 1
Patch Set 1:

... Glad you asked. The answer is that we won't have to create useless defines as much in the Haiku source code itself (https://github.com/haiku/haiku/blob/8f16317a5b6db5c672f331814273e5857555020f/src/libs/libunwind/setjmp/siglongjmp.c#L35) OR depend on the software itself, which could even provide inconsistent values for _NSIG.

We do not want to depend on the definitions by software that was not designed with Haiku in mind, despite the fact that the software could otherwise be fully compatible, particularly because various implementations of _NSIG define values ranging from 64, 65, to 128 and even 129. (e.g. Linux on MIPS is special because it can have 128 signals: https://docs.huihoo.com/doxygen/linux/kernel/3.7/arch_2mips_2include_2uapi_2asm_2signal_8h_source.html)

Some other software may not even provide a macro that either manually defines the arithmetic value of _NSIG or even provide a macro that makes _NSIG equal to NSIG.

Porting such software to Haiku means that we will have to unnecessarily bother with patches. This isn't a hypothetical scenario: As a matter of fact, we already do this on one particular occasion where we are maintaining a patch for screen simply because we do not have this implemented in Haiku: https://github.com/haikuports/haikuports/blob/master/app-misc/screen/patches/screen-4.99.0~git.patchset

A few notable examples of software that better illustrate the inconsistency that we definitely want to stay away from have to be gdb (which supports both NSIG and _NSIG), whereas lvm2 only supports _NSIG, which is not a problem because _NSIG is defined in FreeBSD in the same exact way as the way I did this over here, so nobody really cares. (Source: https://github.com/lvmteam/lvm2/blob/master/daemons/cmirrord/clogd.c#L150)

Do we really want to be at the mercy of the authors and expect that everyone will follow the route of, say, gdb, when they rightfully don't care whatsoever simply because their software runs just fine on FreeBSD?

FreeBSD clearly thinks it's useful enough to include and is enjoying several advantages that we don't, why should we torture ourselves for no reason whatsoever? This change can clearly affect software porting positively, especially as far as NetBSD/OpenBSD-specific software is concerned - I'll reiterate that some of that software can only run on FreeBSD simply because FreeBSD has this particular change that I'm currently proposing.

Patch-set: 1