Update patch set 1
Patch Set 1:

> So, please document things in commit messages.

Acknowledged. Not sure if I should, considering that this doesn't seem likely to get implemented. I just assumed that this wouldn't be a problem in the same sense that nobody really explained themselves.

> > Porting such software to Haiku means that we will have to unnecessarily bother with patches. This isn't a hypothetical scenario: As a matter of fact, we already do this on one particular occasion where we are maintaining a patch for screen simply because we do not have this implemented in Haiku: https://github.com/haikuports/haikuports/blob/master/app-misc/screen/patches/screen-4.99.0~git.patchset

> In the case of screen this was fixed in 2016: https://git.savannah.gnu.org/cgit/screen.git/commit/src/attacher.c?id=c0ecf3bcd24e6bce46440a1fea8ed4d03afb7801
> 
> And in 2017 that function was completely removed: https://git.savannah.gnu.org/cgit/screen.git/commit/src/attacher.c?id=5005299b77ed2815337db1cf58f9b9fb6802dc03

Acknowledged. But this has still happened before, and our range of ~2000 pieces of ported software is wide enough to cover a lot of edge cases, but not a lot of them. Especially with a lot of toolchains and lower-level software.

> Wouldn't such software also be broken on FreeBSD, Linux, and essentially anything that is not OpenBSD? That seems quite unlikely.

It wouldn't be broken on FreeBSD, because they have the change that I tried to implement. We don't. So, it's like we are arbitrarily limiting the software that can run by default on Haiku.  

> They should fix their code which is using nonstandard features without checking if they are available. We are not at their mercy, they are at ours. We write the OS and they have to patch their apps to run on it the correct way.

They probably won't. As I said, most look at FreeBSD, where it runs, because they define the variable in their compatibility layer. I would probably write such software as well if I weren't aware or mindful of Haiku's quirk in that regard, which most likely is the case as far as software that's being written for *BSD's specifically.

Also, excuse my relatively cynic mindset, but every time I think of a piece of software that needs to be supported on Haiku, I expect that the onus is on HaikuPorts contributors by default unless if otherwise stated.

I seriously have no idea what we have to lose and why we are headed towards a conscious decision of borking software for the sake of one extra macro, considering that there's basically no foreseeable chance of this being a slippery slope from my point of view. If it can run on FreeBSD, shouldn't our compatibility layer which is pretty much meant to run as many things as possible that can run on BSD kernels run the software that can run on BSD? Is our compatibility layer strictly limited to FreeBSD?

> Also note that POSIX decided to not introduce this constant because there is no reason for the signal values to be contiguous and positive. What if you have a signal whose value is -1? or 4096 (even if there are no signals at values between 127 and 4095)? So, in fact it turns out NSIG doesn't even make sense and code should not be written to rely on it.

I agree that it doesn't make sense. Which is why I used the assert function in my code here (https://review.haiku-os.org/c/haiku/+/4126) which deliberately breaks programs using the readpassphrase(3) function if they use or depend on signals that basically are out of range, for example on an extremely hypothetical scenario. That's what I define as broken and nonsensical to the point where patches are needed, because this sort of thing can actually cause undefined behavior. I strongly believe that that's where the line should be. If the software can still work perfectly fine as-is, even if it's annoying enough to deal with, then what sort of message are we supposed to send with this?

These programs literally check if _NSIG is not defined (a direct equivalent of it is) then define an arbitrary number, the number usually being "as many signals as I can feasibly imagine in every system whatsoever", which is completely irrelevant to our needs.

By the same exact logic, we should just remove NSIG entirely, because it's useless and annoying to deal with. Or... just go through with this the whole way instead of leaving this basically half-done.

Haiku is an amazing platform that can run programs designed for a beautiful but obscure operating system from 20+ years ago, all the way up to Flash applications and all sorts of fancy emulators. It's amazing for software and media conservation and it's unbelievable how we are one of the few non BSD-based operating systems that can run BSD software. Why would we not want to do this and not enhance an already strong point of ours? I really don't get the message we are supposed to send with this.

Patch-set: 1