Update patch set 1
Patch Set 1:

Hi,

I do not have a strong opinion about this particular change (it seems like we already support FreeBSDs broken define, so we might as well do the same for another BSD), but I have voiced reservations in previous similar compatibility changes, and I will do so again. Many of these no-op compatibility changes are done with the motivation to help porters make it easier to build software without patches. While I appreciate that maintaining patches is a lot of work (I do that for Rust and many of its dependencies), we should also be conscious of the fact that porting software does not end with making it build, but that it should also run properly. And often the broken behaviour will not show itself until you start to intensively use the application.

A compiler error helps with identifying when something might not work as expected. It will give a porter the opportunity to review the impact of the missing symbol, and help decide if (1) it should be functionally implemented, (2) it can be safely supplanted with a no-op compatibility fix or (3) we need to maintain patches to work around the missing functionality in our software.

At this stage I am wondering how much functionality in complex products (like the Rust compiler) is broken and leads to odd runtime errors, because of these quick no-op stubs.

These aspects will need to be part of the analysis when introducing any stubs in the future.

Patch-set: 1