Update patch set 2
Patch Set 2:

In this case it is not a stub. It's just that there are 3 defines for the same thing: __MAX_SIGNO in Haiku, NSIG in FreeBSD and glibc, and _NSIG in OpenBSD and Linux kernel. They all do the same thing. It is a bit sad and unfortunate that people could not agree on a single name for this where it is applicable.

This does not create any problems for ported software, the constant is correctly defined with the proper value.

If we want to be strict about this, also, none of these should be defined in headers/posix.

The main concern here is that an application would attempt to #define NSIG or _NSIG on its own, and get an error because it is already defined. At least _NSIG is starting with an underscore, meaning it is normally in the namespace reserved for the OS and apps should not try to define it themselves. So NSIG is the one that's most nonstandard, and should be moved to headers/bsd and/or headers/gnu.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>