Update patch set 2
Patch Set 2:

> The main concern here is that an application would attempt to #define NSIG or _NSIG on its own, and get an error because it is already defined. At least _NSIG is starting with an underscore, meaning it is normally in the namespace reserved for the OS and apps should not try to define it themselves. So NSIG is the one that's most nonstandard, and should be moved to headers/bsd and/or headers/gnu.

That is something I can definitely agree with (particularly with headers/bsd), I just did not want to be quite invasive with my change. But should I also include _NSIG, which is even more non-standard considering that everyone keeps using it in spite of the underscore?

I think I'll have to correct something here for the record: Every application that tries defines NSIG or _NSIG uses #ifndef beforehand, basically always. In some cases, the biggest possible value according to the developer (keep in mind: the value also changes because of e.g. architecture-specific toolchains. We should probably be in full control of that, like Linux is, for example.) is assigned. In some other cases, they just do what I am trying to do. And that makes sense, because if they didn't do the latter, the application would be broken on BSD.

Regarding what Niels said, the introduction of this stub will not break anything (we already have it, just under another name) and can do more good than bad, because applications implementing their own stubs could be more prone to be broken during execution - if there's a problem with the application, that has to be brought up during compilation. If it's not broken and if it's supposed to work as-is, then why break it?

Patch-set: 2