Update patch set 2
Patch Set 2:

> That is something I can definitely agree with (particularly with headers/bsd), I just did not want to be quite invasive with my change. But should I also include _NSIG, which is even more non-standard considering that everyone keeps using it in spite of the underscore?

_NSIG is more correct for us. We are the OS. We should only use constants that start with an underscore for our nonstandard defines.

NSIG is something that belongs to the applications. They are (in theory) free to use that in their code because we are not supposed to define it.

> 
> I think I'll have to correct something here for the record: Every application that tries defines NSIG or _NSIG uses #ifndef beforehand, basically always. 

1) You cannot proove this. There are many applications
2) It doesn't matter. Let's try to be technically correct if we can.


> Regarding what Niels said, the introduction of this stub will not break anything (we already have it, just under another name) 

This is not a stub. Please let's not use names improperly, it makes things even more confusing than they already are.


So:

- In headers/posix: we are allowed to define _NSIG. We are not allowed to define NSIG (by doing so, we are in fact not POSIX compliant)
- In headers/bsd or headers/gnu: we can define NSIG there if we want to (protected by _BSD_SOURCE, _GNU_SOURCE or _DEFAULT_SOURCE)

Here is the section of POSIX that specifies what is allowed and what isn't: https://pubs.opengroup.org/onlinepubs/9699919799/functions/V2_chap02.html

In particular:

"Additional symbols not required or explicitly permitted by POSIX.1-2017 to be in that header shall not be made visible, except when enabled by another feature test macro."

"The prefixes posix_, POSIX_, and _POSIX_ are reserved for use by POSIX.1-2017 and other POSIX standards. Implementations may add symbols to the headers shown in the following table, provided the identifiers for those symbols either:

Begin with the corresponding reserved prefixes in the table, or

Have one of the corresponding complete names in the table, or

End in the string indicated as a reserved suffix in the table and do not use the reserved prefixes posix_, POSIX_, or _POSIX_, as long as the reserved suffix is in that part of the name considered significant by the implementation.

Symbols that use the reserved prefix _POSIX_ may be made visible by implementations in any header defined by POSIX.1-2017."

"<signal.h>: sa_, si_, sigev_, sival_, uc_, BUS_, CLD_, FPE_, ILL_, SA_, SEGV_, SI_, SIGEV_,"

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>