Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> Is it worth adding a test that demonstrates the broken and fixed behaviour?

It would be nice, but it requires some development in the Python http server used for tests, which is currently not able to generate correct replies to HEAD, or generate a reply with the 204 status and no content.

Right now I am mainly interested in fixing the most annoying issues in haikuwebkit for the beta3 release and I'm not sure I will have time to dig into that.

Patch-set: 1