Update patch set 3
Patch Set 3:

> There was also code that preferred the disk that contained the currently loaded haiku_loader.efi image, but that was removed by tqh's giant refactor, IIRC.

I'm not that fond of naming it like that, it was never intended to be merged as is, and the big problem was that another change added an extra partition to the anyboot. I had more work to fix things, but the whole situation got inflamed and I lost energy. So for me it would be nice if we stop the naming and sometimes shaming.

Patch-set: 3