Update patch set 2
Patch Set 2:

We already got the list of partitions from boot platform before this change, so I'm not sure it should have been applied as is. The part that was missing was picking the right partition to boot from and that can be done in devices.cpp with iterating partitions from UEFI. It would be better if the list of partitions from VFS had more metadata though.

It might be that one of the problems was that bootable partitions started returning non BFS partitions, which meant that if it was the first one, you'd end up in boot menu.

Now we make booting more complex on all platforms, which is not ideal if disk access or mounting is slow.

So what is the plan going forward?

Patch-set: 2