Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Since the API is supposed to become public one day, why not just add a "using BPrivate::..." instead?

Currently, the legacy version of the API (libbnetapi) still exists outside the namespace. I am not confident in importing the namespace as there is a risk of conflicting symbols between the two.

We can change this when the libbnetapi version of the API is removed (we plan to do this in beta4 after everyone had time to get beta3 and migrate their apps to the new private API). For now it's less confusing to explicitly use the namespace.

Patch-set: 4