Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What's the purpose of these extensions? To trim data on a device directly, bypassing the file system? If you're using this on a device with a file system, I'm not sure this makes that much sense.

Prior to this patch (assuming the range size was set correctly to the max. possible value of the used data type instead of letting it overflow to -1), fstrim did not differentiate between a block device and a file system. If a block device was specified, it was completely erased. If a file system was specified, it was correctly trimmed. So I added the --discard-device parameter to avoid accidentally erasing the whole drive.

The offset and length parameters are great for debugging purposes, but they also serve the purpose of trimming only specific device blocks (when run on the device) or only specific parts of the file system (when run on a file system). Apparently, it makes sense to have an ability to specify the offset and length even when trimming a file system. The BFS code does not implement this yet, but there is a "// TODO: take given offset and size into account!" implying that it should be supported. Moreover, fstrim on Linux allows the specification of the offset and the length.

Patch-set: 1