Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (4 comments)
> 
> Thanks for the patch! This seems to mostly fix issues introduced by the change to off_t for the trim data. I don't understand why this was done. Adrien, could you explain this? AFAICT it only makes the code more complex, but I don't see that it has any advantages.

I made this change when adding a call to translate_partition_access() in the devfs so that trim requests would be properly translated between partition and disk offsets (just like read and write calls). This function uses off_t parameters, so viewed from there, it seemed simpler to use off_t rather than have a cast before calling that function.

In devfs, all disk offsets are in off_t (not only for trim, but also for read, write, etc).

You can see that in the BFS code, off_t is also used for CheckBlocks (the interface to checkfs).

So it also seemed to make sense from a consistency point of view.

If bfs uses uint64 and devfs uses off_t, the conversion and bounds checks will have to happen somewhere between the two. I don't know this code well enough to decide where is the best place.

Patch-set: 2