Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What's the purpose of these extensions? To trim data on a device directly, bypassing the file system? If you're using this on a device with a file system, I'm not sure this makes that much sense.

The low level support for directly trimming a device from userspace is useful. For example, this could be used when formatting a partition, to fully reset it before writing anything to it. This is done by mkfs.f2fs on Linux, for example (f2fs is a filesystem specifically designed for use on flash storage like SD cards).

And it seems a good idea to have this protected by some command line flags in the tool. It was indeed a source of confusion for some people testing the trim support, who accidentally erased a whole partition instead of trimming it.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>