Update patch set 8
Patch Set 8:

This update includs:
- Checking the LaunchRoster initialization
- Using the Application Signature to get the JobInfo instead of checking each job

However, there is a problem with two apps. These apps have different signatures than what the LaunchRoster expects:

LaunchRoster / Application
x-vnd.haiku-media_server / x-vnd.Be.media-server
x-vnd.haiku-power_daemon / x-vnd.Haiku-powermanagement

The TeamID of does match, so it is a signature mismatch.  The easy fix looks like updating the *.rdef files. Is this something that should be fixed in this patch or a separate bug?

Patch-set: 8