Update patch set 1
Patch Set 1:

Ideally, the Haiku kernel should use the background image of that release, not the background embedded in the haiku_loader, but that would require either zlib in the kernel, which is probably why the current method used; or perhaps embedding a resource in the haiku kernel that the loader could read and load.

But this helps with ensuring the row of icons is aligned no matter which haiku + loader combination is used together.

Patch-set: 1