Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Should I also update dates in copyright headers? If so, which ones and how?

You can either add yourself with a date of 2021 if you want to keep your copyright, or you can reassign the copyright to Haiku inc, in that case change the date for Haiku inc to 2021. If you made only small changes, it is not really required to update it. However, it is not clear when a change is small, as this is a bit subjective (changing a type with another I would consider "small", even if it ends up touching a lot of lines in the file).

Wether you can actually reassign copyright depends on local laws and we can't say for you.

If we ever need to ask your permission for a license change, in addition to the copyright headers, we will also check the "git blame" or similar tools to collect a more complete list of authors.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>