Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I remember somehow that the cross tools are moved around after build (I think kallisti5 mentioned this). The rpath wouldn't be correct anymore, or this doesn't matter?

This change is not related to the cross tools. What it does is add to all things built for the host (all things built from src/build/) an rpath to generated/objects/host/.../ so that the relevant .so files can be found, without need to mess with LD_LIBRARY_PATH or using "jam run".

If the rpath is invalid, the existing methods (LD_LIBRARY_PATH, jam run) can be used as a fallback.

So, the only thing that changes here, is that, thanks to the rpath, these tools can now be run directly from the command line, additionally to the other ways they are already run by the build system

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>