Fix disappearing text in WebKit.
The ClipToRect implementation in BView attempts to optimize clipping by
using a BRegion rather than a shape/picture when there is no transform.
It also does this when the transform is only scaling and translation
(the result should still be a rectangle). For some reason, the latter
case is not working properly, and this results in disappearing text in
WebKit.

Fixes #17066