Update patch set 2
Patch Set 2:

(4 comments)

> Patch Set 2:
> 
> ```
> Starting build of type regular ... 
> Building Haiku Nightly 
> Asked for bios_ia32 target boot platform 
> Unknown path to handle adding to image 
> Asked for pxe_ia32 target boot platform 
> Unknown path to handle adding to image 
> don't know how to make @nightly-anyboot
> ...patience...
> ...patience...
> ...found 14381 target(s)...
> ...can't find 1 target(s)...
> ```
> 
> Sorry about this, accidentally hit CTRL + Enter as I was typing this.

That error is normal and is due to the hacky way we implemented build profiles.

When you specify both a build profile and some targets, jam is a bit confused and tries to use the build profile also as a target.

The reason it doesn't build anything is simply because the targets you asked for are already built.

Patch-set: 2